requirements.txt
contourpy==1.2.0
cycler==0.12.1
fonttools==4.46.0
kiwisolver==1.4.5
matplotlib==3.8.2
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.2
Pillow==9.4.0
pyparsing==3.1.1
python-dateutil==2.8.2
six==1.16.0
